fix(archive/chat): rework live chat to tdl chat parsing
Attempt to correct invalid emote positions from chat-downloader
